requirements.txt
requests
pandas
openpyxl
python-dateutil
tqdm
PyYAML
PyQt6
